movie_rag_requirements.txt
# Core LangChain and OpenAI
langchain
langchain_openai
openai
python-dotenv

# Vector Database
chromadb

# Web Scraping
selenium
beautifulsoup4
requests

# Data Processing
pandas
numpy

# Additional utilities
tqdm 